Adding Searching Raports by Date and Type

+++ ASPNetCore/CEIDGASPNetCore/DbModel/RaportTypeNames.cs
@@ -6,7 +6,7 @@ namespace CEIDGASPNetCore.DbModel
     {
         [Key]
         public long Id { get; set; }
-        public int RaportType { get; set; }
+        public byte RaportType { get; set; }
 
         public string RaportTypeName { get; set; }
     }
